Keep supported_elements like before merging
