+++ geos-mesh/src/geos/mesh/doctor/actions/supported_elements.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 import multiprocessing
 import networkx
+from numpy import ones
 from tqdm import tqdm
 from typing import FrozenSet, Iterable, Mapping, Optional
 from vtkmodules.util.numpy_support import vtk_to_numpy
@@ -29,40 +30,24 @@ class Result:
 
 # for multiprocessing, vtkUnstructuredGrid cannot be pickled. Let's use a global variable instead.
 MESH: Optional[ vtkUnstructuredGrid ] = None
-
-
-def init_worker_mesh( input_file_for_worker: str ):
-    """Initializer for multiprocessing.Pool to set the global MESH variable in each worker process.
-
-    Args:
-        input_file_for_worker (str): Filepath to vtk grid
-    """
-    global MESH
-    setup_logger.debug(
-        f"Worker process (PID: {multiprocessing.current_process().pid}) initializing MESH from file: {input_file_for_worker}"
-    )
-    MESH = read_mesh( input_file_for_worker )
-    if MESH is None:
-        setup_logger.error(
-            f"Worker process (PID: {multiprocessing.current_process().pid}) failed to load mesh from {input_file_for_worker}"
-        )
-        # You might want to raise an error here or ensure MESH being None is handled downstream
-        # For now, the assert MESH is not None in __call__ will catch this.
+supported_cell_types: set[ int ] = {
+    VTK_HEXAGONAL_PRISM, VTK_HEXAHEDRON, VTK_PENTAGONAL_PRISM, VTK_POLYHEDRON, VTK_PYRAMID, VTK_TETRA, VTK_VOXEL,
+    VTK_WEDGE
+}
 
 
 class IsPolyhedronConvertible:
 
-    def __init__( self ):
+    def __init__( self, mesh: vtkUnstructuredGrid ):
+        global MESH  # for multiprocessing, vtkUnstructuredGrid cannot be pickled. Let's use a global variable instead.
+        MESH = mesh
 
         def build_prism_graph( n: int, name: str ) -> networkx.Graph:
-            """Builds the face to face connectivities (through edges) for prism graphs.
-
-            Args:
-                n (int): The number of nodes of the basis (i.e. the pentagonal prims gets n = 5)
-                name (str): A human-readable name for logging purpose.
-
-            Returns:
-                networkx.Graph: A graph instance.
+            """
+            Builds the face to face connectivities (through edges) for prism graphs.
+            :param n: The number of nodes of the basis (i.e. the pentagonal prims gets n = 5)
+            :param name: A human-readable name for logging purpose.
+            :return: A graph instance.
             """
             tmp = networkx.cycle_graph( n )
             for node in range( n ):
@@ -90,34 +75,26 @@ class IsPolyhedronConvertible:
         }
 
     def __is_polyhedron_supported( self, face_stream ) -> str:
-        """Checks if a polyhedron can be converted into a supported cell.
+        """
+        Checks if a polyhedron can be converted into a supported cell.
         If so, returns the name of the type. If not, the returned name will be empty.
-
-        Args:
-            face_stream (_type_): The polyhedron.
-
-        Returns:
-            str: The name of the supported type or an empty string.
+        :param face_stream: The polyhedron.
+        :return: The name of the supported type or an empty string.
         """
         cell_graph = build_face_to_face_connectivity_through_edges( face_stream, add_compatibility=True )
-        if cell_graph.order() not in self.__reference_graphs:
-            return ""
         for reference_graph in self.__reference_graphs[ cell_graph.order() ]:
             if networkx.is_isomorphic( reference_graph, cell_graph ):
                 return str( reference_graph.name )
         return ""
 
     def __call__( self, ic: int ) -> int:
-        """Checks if a vtk polyhedron cell can be converted into a supported GEOSX element.
-
-        Args:
-            ic (int): The index element.
-
-        Returns:
-            int: -1 if the polyhedron vtk element can be converted into a supported element type. The index otherwise.
+        """
+        Checks if a vtk polyhedron cell can be converted into a supported GEOSX element.
+        :param ic: The index element.
+        :return: -1 if the polyhedron vtk element can be converted into a supported element type. The index otherwise.
         """
         global MESH
-        assert MESH is not None, f"MESH global variable not initialized in worker process (PID: {multiprocessing.current_process().pid}). This should have been set by init_worker_mesh."
+        assert MESH is not None
         if MESH.GetCellType( ic ) != VTK_POLYHEDRON:
             return -1
         pt_ids = vtkIdList()
@@ -128,50 +105,40 @@ class IsPolyhedronConvertible:
             setup_logger.debug( f"Polyhedron cell {ic} can be converted into \"{converted_type_name}\"" )
             return -1
         else:
-            setup_logger.debug(
-                f"Polyhedron cell {ic} (in PID {multiprocessing.current_process().pid}) cannot be converted into any supported element."
-            )
+            setup_logger.debug( f"Polyhedron cell {ic} cannot be converted into any supported element." )
             return ic
 
 
-def __action( vtk_input_file: str, options: Options ) -> Result:
-    # Main process loads the mesh for its own use
-    mesh = read_mesh( vtk_input_file )
-    if mesh is None:
-        setup_logger.error( f"Main process failed to load mesh from {vtk_input_file}. Aborting." )
-        # Return an empty/error result or raise an exception
-        return Result( unsupported_std_elements_types=frozenset(), unsupported_polyhedron_elements=frozenset() )
-
-    if hasattr( mesh, "GetDistinctCellTypesArray" ):
-        cell_types_numpy = vtk_to_numpy( mesh.GetDistinctCellTypesArray() )
-        cell_types = set( cell_types_numpy.tolist() )
+def find_unsupported_std_elements_types( mesh: vtkUnstructuredGrid ) -> set[ int ]:
+    if hasattr( mesh, "GetDistinctCellTypesArray" ):  # For more recent versions of vtk.
+        unique_cell_types = set( vtk_to_numpy( mesh.GetDistinctCellTypesArray() ) )
     else:
-        vtk_cell_types_obj = vtkCellTypes()
-        mesh.GetCellTypes( vtk_cell_types_obj )
-        cell_types = set( vtk_iter( vtk_cell_types_obj ) )
+        vtk_cell_types = vtkCellTypes()
+        mesh.GetCellTypes( vtk_cell_types )
+        unique_cell_types = set( vtk_iter( vtk_cell_types ) )
+    return unique_cell_types - supported_cell_types
 
-    supported_cell_types = {
-        VTK_HEXAGONAL_PRISM, VTK_HEXAHEDRON, VTK_PENTAGONAL_PRISM, VTK_POLYHEDRON, VTK_PYRAMID, VTK_TETRA, VTK_VOXEL,
-        VTK_WEDGE
-    }
-    unsupported_std_elements_types = cell_types - supported_cell_types
 
+def find_unsupported_polyhedron_elements( mesh: vtkUnstructuredGrid, options: Options ) -> list[ int ]:
     # Dealing with polyhedron elements.
-    num_cells = mesh.GetNumberOfCells()
-    polyhedron_converter = IsPolyhedronConvertible()
+    num_cells: int = mesh.GetNumberOfCells()
+    result = ones( num_cells, dtype=int ) * -1
+    with multiprocessing.Pool( processes=options.num_proc ) as pool:
+        generator = pool.imap_unordered( IsPolyhedronConvertible( mesh ),
+                                         range( num_cells ),
+                                         chunksize=options.chunk_size )
+        for i, val in enumerate( tqdm( generator, total=num_cells, desc="Testing support for elements" ) ):
+            result[ i ] = val
+    return [ i for i in result if i > -1 ]
 
-    unsupported_polyhedron_indices = []
-    # Pass the vtk_input_file to the initializer
-    with multiprocessing.Pool( processes=options.nproc, initializer=init_worker_mesh,
-                               initargs=( vtk_input_file, ) ) as pool:  # Comma makes it a tuple
-        generator = pool.imap_unordered( polyhedron_converter, range( num_cells ), chunksize=options.chunk_size )
-        for cell_index_or_neg_one in tqdm( generator, total=num_cells, desc="Testing support for elements" ):
-            if cell_index_or_neg_one != -1:
-                unsupported_polyhedron_indices.append( cell_index_or_neg_one )
 
+def __action( mesh: vtkUnstructuredGrid, options: Options ) -> Result:
+    unsupported_std_elements_types: set[ int ] = find_unsupported_std_elements_types( mesh )
+    unsupported_polyhedron_elements: list[ int ] = find_unsupported_polyhedron_elements( mesh, options )
     return Result( unsupported_std_elements_types=frozenset( unsupported_std_elements_types ),
-                   unsupported_polyhedron_elements=frozenset( unsupported_polyhedron_indices ) )
+                   unsupported_polyhedron_elements=frozenset( unsupported_polyhedron_elements ) )
 
 
 def action( vtk_input_file: str, options: Options ) -> Result:
-    return __action( vtk_input_file, options )
+    mesh: vtkUnstructuredGrid = read_mesh( vtk_input_file )
+    return __action( mesh, options )
